FMCG/requirements.txt
pandas
faker
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
pandas_gbq
